Broken windows and building models
Added broken window sprites to the windows when they are in destruction range of an exploding banana.

Added 3 windows to the 2-width building and 4 to the 3-width to stop them looking off-scale.

Tried to add the ability to show/hide the trajectory line but that didnt work. Will try again