AI attempt 2 (fail) so rewrite of throwing and trajectory line code
Fun one tonight.  All of the calculations for the power done by the AI give the result as a force, not a velocity (which does make sense) so that required the rewriting of the launching of the projectile and the calculations that go into the trajectory line.   Big hassle, but done now (i hope).

AI attempt 3 will follow tomorrow