First powerup
Added powerup code to the game in the form of buttons that are created when a powerup is hit.  Added the trajectory line button as the first powerup.

Fixed a bug where the UI for a player was being created each round.
Stopped the player getting destroyed after being hit; now it gets deactivated